chg: [website] WIP - A new Flask blueprint dedicated to the users has been added with another blueprint dedicated to the admin views. These blueprints are optional.
It is possible to create admin users thanks to a Flask script. These users are
stored in a dedicated database. Permissions: WIP.

The default views are now in a blueprint named 'home' and the API has
as well a dedicated blueprint. No changes for the clients.

The architecture of the Flask app in the 'website' folter has been
slightly restructured for more clarity but there is no change for
the users (endpoints are the same). It will be easier to scale the
app in the future. WIP.